implement add api command
