+++ components/micro-gateway-cli/src/main/java/org/wso2/apimgt/gateway/cli/cmd/GatewayCliCommands.java
@@ -25,4 +25,5 @@ public class GatewayCliCommands {
     public static final String BUILD = "build";
     public static final String RUN = "run";
     public static final String RESET = "reset";
+    public static final String ADD = "add";
 }
